Add 'overlap' reward for state preparation with universal gate set.
This reward is the overlap with the target state at the end of the
episode. It's just an intermediate thing, will need to construct a
sample-based approximation to this to work in real experiment.

The target state is specified by the user and passed as a parameter
to GKP class.