Start transition to new simulator API. This breaks compatibility.
This will be followed by another commit that fixes up some of the
code to make sure SNAP circuit can run without errors, and then I
will patch up the rest of the code as things come up. Probably I
should have done this in a separate branch, but who cares.

This commit:
* New simulator/operators.py with LinearOperator API.
* New simulator/utils.py