Multiple fixes to make things consistent with new simulator API.
Found a big problem with the new API: tensorflow treats each new
object of LinearOperator class as a separate node in the graph.
This means that the memory usage keeps growing, and eventually it
reaches the memmory allocation error. I didn't anticipate this in
advance, but it means that this API will not be useful. Will revert
to the older commit and open a new branch to develop a reasonable
simulator API which treats operators as simple tensors.