Replace most of the operators in OscillatorQubit using new simulator API.
This uses conversion to LinearOperator to calculate the tensor products.
I checked that this time it doesn't clog the GPU memory, and things run
as fast as they used to run with the BCH mixin.

1000 epochs trainining for state prep with SNAPs with old mixin class takes
7.93 mins, with a new SNAP using kronecker products it takes 8.36 mins, and
with new SNAP and new displacement (both using Kronecker products) it's 7.6
mins. So it seems that it doesn't hurt the simulation time in any noticeable
way.